+++ src/app/core/components/guild-selector/guild-selector.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-guild-selector',
+  imports: [],
+  templateUrl: './guild-selector.component.html',
+  styleUrl: './guild-selector.component.css'
+})
+export class GuildSelectorComponent {
+
+}