+++ src/app/guild-config/guild-config.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-guild-config',
+  imports: [],
+  templateUrl: './guild-config.component.html',
+  styleUrl: './guild-config.component.css'
+})
+export class GuildConfigComponent {
+
+}